requirements.txt
fastmcp>=2.8.0
python-dotenv>=1.1.0
requests>=2.31.0
openai>=1.0.0
boto3>=1.38.0
yfinance>=0.2.60
requests-cache>=1.2.0
pandas>=2.3.0
numpy>=2.3.0
snowflake-connector-python>=3.15.0
cryptography>=45.0.0 